楼区管理员只能看本楼区和中心库房的设备
