+++ src/app/asset/area/list/checkout.dialog.html
@@ -1,7 +1,7 @@
-<modal-dialog modalTitle="领用单">
+<modal-dialog modalTitle="{{transferType}}单">
     <modal-body>
         <form class="form-horizontal">
-            <form-field label="领用楼区">
+            <form-field label="{{transferType}}楼区">
                 <select name="building" class="form-control" [(ngModel)]="form.building" (ngModelChange)="buildingChange($event)">
                     <option *ngFor="let building of buildings " [ngValue]="building">
                         {{ building}}
@@ -15,8 +15,8 @@
                     </option>
                 </select>
             </form-field>
-            <form-field label="领用说明">
-                <input class="form-control" name="note" type="text" [(ngModel)]="form.note" placeholder="领用说明、备注" />
+            <form-field label="{{transferType}}说明">
+                <input class="form-control" name="note" type="text" [(ngModel)]="form.note" placeholder="{{transferType}}说明、备注" />
             </form-field>
         </form>
         <tm-asset-device-grid [data]="assets" ></tm-asset-device-grid>
