+++ src/app/asset/area/list/form-list.component.html
@@ -2,6 +2,7 @@
     <div class="btn-group ml-auto">
         <button class="btn btn-secondary" (click)="query()">查询</button>
         <button class="btn btn-secondary ml-2" (click)="checkout()">领用</button>
+        <button class="btn btn-secondary ml-2" routerLink="../forms">流转单</button>
     </div>
 </div>
 <tm-asset-device-grid [data]="assets" [checkboxSelection]="true" [flush]="flush" (rowSelected)="onRowSelected($event)">
