+++ src/app/asset/area/shared/form-viewer.component.html
@@ -26,7 +26,7 @@
     </div>
     <div class="row">
         <div class="form-group col-12">
-            <label class="control-label">入库单明细</label>
+            <label class="control-label">{{vm.type}}单明细</label>
             <tm-asset-device-grid [data]="vm.items"></tm-asset-device-grid>
         </div>
     </div>
