+++ src/app/asset/area/shared/form.model.ts
@@ -1,39 +1,6 @@
+import { Asset } from '../../shared/asset-form.model';
+
 export const statusLabel = { CREATED: '未提交', SUBMITTED: '提交', APPROVED: '同意', REJECTED: '不同意'};
-export class Asset {
-    id: number;
-    sn: string;
-    code: string;
-    price: number;
-    dateBought: string;
-    qualifyMonth: number;
-    dateForbid: string;
-    dateClose: string;
-    state: string;
-    fault: boolean;
-    supplierId: number;
-    assetModelId: number;
-    assetType: string;
-    name: string;
-    brand: string;
-    specs: string;
-    parameter: string;
-    unit: string;
-    pcs: number;
-    total: number;
-    supplier: string;
-    note: string;
-
-    constructor(dto: any) {
-        Object.assign(this, dto);
-    }
-
-    equalsTo(other: Asset): boolean {
-        if (this.id && other.id && this.id === other.id) {
-            return true;
-        }
-        return false;
-    }
-}
 
 export class TransferForm {
     id: number;
